refactor(msteams): isolate message admission
