+++ extensions/msteams/src/monitor-handler/access.ts
@@ -1,4 +1,6 @@
 // Msteams plugin module implements access behavior.
+import { formatAllowlistMatchMeta } from "openclaw/plugin-sdk/allow-from";
+import { logInboundDrop } from "openclaw/plugin-sdk/channel-inbound";
 import {
   channelIngressRoutes,
   resolveStableChannelMessageIngress,
@@ -12,8 +14,12 @@ import {
   resolveDefaultGroupPolicy,
   type OpenClawConfig,
 } from "../../runtime-api.js";
+import type { StoredConversationReference } from "../conversation-store.js";
+import type { MSTeamsConversationStore } from "../conversation-store.js";
+import { formatUnknownError } from "../errors.js";
 import { normalizeMSTeamsConversationId } from "../inbound.js";
-import { resolveMSTeamsRouteConfig } from "../policy.js";
+import type { MSTeamsMonitorLogger } from "../monitor-types.js";
+import { resolveMSTeamsAllowlistMatch, resolveMSTeamsRouteConfig } from "../policy.js";
 import { looksLikeMSTeamsConversationId } from "../resolve-allowlist.js";
 import { getMSTeamsRuntime } from "../runtime.js";
 import type { MSTeamsTurnContext } from "../sdk-types.js";
@@ -68,6 +74,38 @@ function normalizeAllowlistConversationId(value?: string | null): string | null 
   const trimmed = value?.trim();
   // Microsoft Graph conversation IDs are opaque; case-folding would authorize a different chat.
   return trimmed ? normalizeMSTeamsConversationId(trimmed) : null;
+}
+
+function formatMSTeamsSenderReason(params: {
+  reasonCode: string;
+  dmPolicy?: string;
+  groupPolicy?: string;
+}): string {
+  switch (params.reasonCode) {
+    case "dm_policy_open":
+      return "dmPolicy=open";
+    case "dm_policy_disabled":
+      return "dmPolicy=disabled";
+    case "dm_policy_pairing_required":
+      return "dmPolicy=pairing (not allowlisted)";
+    case "dm_policy_allowlisted":
+      return `dmPolicy=${params.dmPolicy ?? "allowlist"} (allowlisted)`;
+    case "dm_policy_not_allowlisted":
+      return `dmPolicy=${params.dmPolicy ?? "allowlist"} (not allowlisted)`;
+    case "group_policy_disabled":
+      return "groupPolicy=disabled";
+    case "group_policy_empty_allowlist":
+    case "route_sender_empty":
+      return "groupPolicy=allowlist (empty allowlist)";
+    case "group_policy_not_allowlisted":
+      return "groupPolicy=allowlist (not allowlisted)";
+    case "group_policy_open":
+      return "groupPolicy=open";
+    case "group_policy_allowed":
+      return `groupPolicy=${params.groupPolicy ?? "allowlist"}`;
+    default:
+      return params.reasonCode;
+  }
 }
 
 export async function resolveMSTeamsSenderAccess(params: {
@@ -167,5 +205,188 @@ export async function resolveMSTeamsSenderAccess(params: {
     channelGate,
     allowNameMatching,
     groupPolicy,
+  };
+}
+
+export async function admitMSTeamsMessage(params: {
+  cfg: OpenClawConfig;
+  activity: MSTeamsTurnContext["activity"];
+  text: string;
+  conversationId: string;
+  conversationRef: StoredConversationReference;
+  isChannel: boolean;
+  conversationStore: MSTeamsConversationStore;
+  log: MSTeamsMonitorLogger;
+  logVerboseMessage: (message: string) => void;
+}) {
+  const core = getMSTeamsRuntime();
+  const allowTextCommands = core.channel.commands.shouldHandleTextCommands({
+    cfg: params.cfg,
+    surface: "msteams",
+  });
+  const isControlCommand =
+    allowTextCommands && core.channel.commands.isControlCommandMessage(params.text, params.cfg);
+  const access = await resolveMSTeamsSenderAccess({
+    cfg: params.cfg,
+    activity: params.activity,
+    hasControlCommand: isControlCommand,
+  });
+  const {
+    dmPolicy,
+    senderId,
+    senderName,
+    pairing,
+    isDirectMessage,
+    channelGate,
+    senderAccess,
+    commandAccess,
+    allowNameMatching,
+    groupPolicy,
+    msteamsCfg,
+  } = access;
+  const effectiveDmAllowFrom = senderAccess.effectiveAllowFrom;
+  const effectiveGroupAllowFrom = senderAccess.effectiveGroupAllowFrom;
+
+  if (isDirectMessage && msteamsCfg && senderAccess.decision !== "allow") {
+    if (senderAccess.reasonCode === "dm_policy_disabled") {
+      params.log.info("dropping dm (dms disabled)", {
+        sender: senderId,
+        label: senderName,
+      });
+      params.log.debug?.("dropping dm (dms disabled)");
+      return null;
+    }
+    const allowMatch = resolveMSTeamsAllowlistMatch({
+      allowFrom: effectiveDmAllowFrom,
+      senderId,
+      senderName,
+      allowNameMatching,
+    });
+    if (senderAccess.decision === "pairing") {
+      params.conversationStore
+        .upsert(params.conversationId, params.conversationRef)
+        .catch((err: unknown) => {
+          params.log.debug?.("failed to save conversation reference", {
+            error: formatUnknownError(err),
+          });
+        });
+      const request = await pairing.upsertPairingRequest({
+        id: senderId,
+        meta: { name: senderName },
+      });
+      if (request) {
+        params.log.info("msteams pairing request created", {
+          sender: senderId,
+          label: senderName,
+        });
+      }
+    }
+    params.log.debug?.("dropping dm (not allowlisted)", {
+      sender: senderId,
+      label: senderName,
+      allowlistMatch: formatAllowlistMatchMeta(allowMatch),
+    });
+    params.log.info("dropping dm (not allowlisted)", {
+      sender: senderId,
+      label: senderName,
+      dmPolicy,
+      reason: formatMSTeamsSenderReason({
+        reasonCode: senderAccess.reasonCode,
+        dmPolicy,
+        groupPolicy,
+      }),
+      allowlistMatch: formatAllowlistMatchMeta(allowMatch),
+    });
+    return null;
+  }
+
+  if (!isDirectMessage && msteamsCfg) {
+    if (channelGate.allowlistConfigured && !channelGate.allowed) {
+      params.log.info("dropping group message (not in team/channel allowlist)", {
+        conversationId: params.conversationId,
+        teamKey: channelGate.teamKey ?? "none",
+        channelKey: channelGate.channelKey ?? "none",
+        channelMatchKey: channelGate.channelMatchKey ?? "none",
+        channelMatchSource: channelGate.channelMatchSource ?? "none",
+      });
+      params.log.debug?.("dropping group message (not in team/channel allowlist)", {
+        conversationId: params.conversationId,
+        teamKey: channelGate.teamKey ?? "none",
+        channelKey: channelGate.channelKey ?? "none",
+        channelMatchKey: channelGate.channelMatchKey ?? "none",
+        channelMatchSource: channelGate.channelMatchSource ?? "none",
+      });
+      return null;
+    }
+
+    if (!senderAccess.allowed && senderAccess.reasonCode === "group_policy_disabled") {
+      params.log.info("dropping group message (groupPolicy: disabled)", {
+        conversationId: params.conversationId,
+      });
+      params.log.debug?.("dropping group message (groupPolicy: disabled)", {
+        conversationId: params.conversationId,
+      });
+      return null;
+    }
+    if (
+      !senderAccess.allowed &&
+      (senderAccess.reasonCode === "group_policy_empty_allowlist" ||
+        senderAccess.reasonCode === "route_sender_empty")
+    ) {
+      params.log.info("dropping group message (groupPolicy: allowlist, no allowlist)", {
+        conversationId: params.conversationId,
+      });
+      params.log.debug?.("dropping group message (groupPolicy: allowlist, no allowlist)", {
+        conversationId: params.conversationId,
+      });
+      return null;
+    }
+    if (!senderAccess.allowed && senderAccess.reasonCode === "group_policy_not_allowlisted") {
+      const allowMatch = resolveMSTeamsAllowlistMatch({
+        allowFrom: effectiveGroupAllowFrom,
+        senderId,
+        senderName,
+        allowNameMatching,
+      });
+      params.log.debug?.("dropping group message (not in groupAllowFrom)", {
+        sender: senderId,
+        label: senderName,
+        allowlistMatch: formatAllowlistMatchMeta(allowMatch),
+      });
+      params.log.info("dropping group message (not in groupAllowFrom)", {
+        sender: senderId,
+        label: senderName,
+        allowlistMatch: formatAllowlistMatchMeta(allowMatch),
+      });
+      return null;
+    }
+  }
+
+  if (commandAccess.shouldBlockControlCommand) {
+    logInboundDrop({
+      log: params.logVerboseMessage,
+      channel: "msteams",
+      reason: "control command (unauthorized)",
+      target: senderId,
+    });
+    return null;
+  }
+
+  params.conversationStore
+    .upsert(params.conversationId, params.conversationRef)
+    .catch((err: unknown) => {
+      params.log.debug?.("failed to save conversation reference", {
+        error: formatUnknownError(err),
+      });
+    });
+
+  return {
+    ...access,
+    allowTextCommands,
+    isControlCommand,
+    commandAuthorized: commandAccess.requested ? commandAccess.authorized : undefined,
+    effectiveDmAllowFrom,
+    effectiveGroupAllowFrom,
+    isChannel: params.isChannel,
   };
 }
